voicehive-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.3.0
google-cloud-aiplatform==1.38.0
google-cloud-secret-manager==2.16.4
google-cloud-logging==3.8.0
google-cloud-monitoring==2.16.0
requests==2.31.0
mem0ai==0.0.11
twilio==8.10.0
firebase-admin==6.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
psutil==5.9.6
hypothesis==6.88.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-prometheus==1.12.0rc1
prometheus-client==0.19.0
